Moved the Locked type up into lib.rs
