+++ kernel/src/allocator/bump.rs
@@ -1,4 +1,5 @@
-use super::{align_up, Locked};
+use super::align_up;
+use crate::Locked;
 use alloc::alloc::{GlobalAlloc, Layout};
 use core::ptr;
