+++ kernel/src/lib.rs
@@ -42,6 +42,23 @@ pub fn halt_loop() -> ! {
     }
 }
 
+// A wrapper around spin::Mutex to permit trait implementations.
+pub struct Locked<A> {
+    inner: spin::Mutex<A>,
+}
+
+impl<A> Locked<A> {
+    pub const fn new(inner: A) -> Self {
+        Locked {
+            inner: spin::Mutex::new(inner),
+        }
+    }
+
+    pub fn lock(&self) -> spin::MutexGuard<A> {
+        self.inner.lock()
+    }
+}
+
 // Testable represents a test function.
 //
 pub trait Testable {
